api/ocr/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
easyocr==1.7.1
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
Pillow==10.1.0
